fix type promotion problem.

+++ paddlenlp/transformers/funnel/modeling.py
@@ -519,7 +519,7 @@ class FunnelRelMultiheadAttention(nn.Layer):
             positional_attn = _relative_shift_gather(positional_attn, context_len, shift)
 
         if cls_mask is not None:
-            positional_attn *= cls_mask
+            positional_attn *= cls_mask.astype(positional_attn.dtype)
         return positional_attn
 
     def relative_token_type_attention(self, token_type_mat, q_head, cls_mask=None):
